[lich.rbw] avoid stomping exits because of doorknobs
if a rusty doorknob is in a room, and a mundane "go door" wayto fails because of it code changes the exit to "go second door" (or third, fourth. it tries to iterate based on the exit) it changes the map DB. However, it changes the entire room in memory not just the command issued to the game so it persists for the entire session. If the doorknob is them removed, the exit is now broken until the player logs out or reloads the mapdb.

by casting the "dir" var to a string first, we avoid this issue by manipulating a temporary string var and not a reference to the map DB in memory itself.