[infomon.rb] dash space fix for naming keys
Currently keys should be automatically replacing spacing with `_` characters. But in regards to lore skills, this is currently having an issue due to the `.gsub('_-_', '')` removing the underline as shown via:
```ruby
;e echo Infomon._key('sorcerous lore - necromancy')
[exec1: sorcerous_lorenecromancy] 
```
Replacing the previous to be `.gsub('_-_', '_')` instead now corrects this and has the naming to be:
```ruby
;e echo Infomon._key('sorcerous lore - necromancy')
[exec1: sorcerous_lore_necromancy] 
```
Which matches skills.rb module for returning proper values.