[infomon] Refactor ActiveSpell watcher to use callback
When testing recent updates to lich, I noticed that my CPU usages was
significantly higher (ie. redlining multiple CPUs on a modern machine).

After some investigation, it looks like the ActiveSpell watcher spawns a
thread that repeatedly checks a new boolean value in the XMLData class.

I refactored this to instead invoke the ActiveSpell method directly
instead of setting a boolean. This has greatly reduced CPU usage to less
than 1% almost all the time. I understand the motivation for moving this
to its own thread though, so I testing this especially with lots of
scroll and spell updates to make sure it doesn't block processing in the
main lich XML parsing routine.