[gs][xmlparser.rb] bugfix: missing real ID in <nav>
Fix to account for missing real id# in the `<nav>` tag for rooms meant to be hidden/puzzle. Instead use an MD5 hash to create an integer based on room title, description, and exits.

Should allow for mapping of rooms to be unique except for day/night conflicts. Also should allow for Claim module to continue to work in these rooms as well.