[gs][xmlparser.rb] bugfix: Claim module obvious hiding
`@check_obvious_hiding` was ONLY being reset upon actually seeing a true obvious hider in a room. Causing the later check:
```ruby
      elsif @check_obvious_hiding && text_string =~ /obvious signs of someone hiding/
        @room_player_hidden = true
```
To happen outside of room entry only after the `<compass>` tag was sent. This would lead to self-party members causing Claim to not claim a room if somehow you LOOK in the room after entry and see the obvious signs of someone hiding. By moving the reset of `@check_obvious_hiding` to outside of the `if @room_player_hidden` check at the end of the Lich::Claim process, this prevents that behavior.