fix(all): restore Wrayth / Wizard script execution ability
After the `games.rb` rewrite, Wrayth / Wizard script execution was impaired.  This PR creates a `JumpError` class to reimplement that execution, and inherits from `StandardError` rather than a simple generic `Exception` to align with modern Ruby practice.  The definition is only used in scripts, so it is moved from the root `lich.rbw` to `script.rb`.  `global_defs` call is updated to point to the new class.