Update main.rb
Added helpers to determine the correct pid for terminal fe when using detached mode. 

Profanity sends Process.pid to lich. Lich uses that pid to traverse the process thread until it finds a process with a visible window and attaches uses that pid as the frontend_pid.

Frostbite would need to send a pid, as well as any other detached mode clients.