fix(gs): match READY/STOW items with a/an/some inside the link (#1026)
Due to inconsistent item creation within game, sometimes the `a`, `an`,
and `some` can be misplaced within the EXIST link. This makes that
optional so that it can be in either location.
<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Updated regex patterns in `xmlparser.rb` to handle optional placement
of 'a', 'an', and 'some' in STOW and READY item links.
> 
>   - **Behavior**:
> - Updated regex patterns in `xmlparser.rb` to allow optional placement
of `a`, `an`, and `some` in item links for STOW and READY lists.
> - Handles cases where these articles appear before or after the `<a
exist>` link.
>   - **Patterns**:
> - Modified `StowListContainer`, `StowSetContainer1`,
`StowSetContainer2` to support optional articles.
> - Modified `ReadyListNormal`, `ReadyListAmmo2`, `ReadyListSheathsSet`,
`ReadyItemSet` to support optional articles.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis"
src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=elanthia-online%2Flich-5&utm_source=github&utm_medium=referral)<sup>
for 247a658fb933d721f2f8f2cd25641629d667f427. You can
[customize](https://app.ellipsis.dev/elanthia-online/settings/summaries)
this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->