Update cleanup strategy for old instances
When the creature registry hits it's max (1000) it does an aggressive age based cleanup to make room. This was set to clear out anything more than 15 minutes old which was way too aggressive when it takes 2+ hours to hit 1000 creatures in constant hunting barring ascension swarms.

Adjusted it to remove entries starting older than 2 hours and getting more aggressive if needed.